+++ packages/react/src/hooks/useIsomorphicLayoutEffect/index.ts
@@ -1,4 +1,4 @@
-import { isClient } from '@devgrace/utils';
+import { isClient } from '@modern-kit/utils';
 import { useEffect, useLayoutEffect } from 'react';
 
 export const useIsomorphicLayoutEffect = isClient()